Move implementation of policies to container/detail
